+++ src/podmander-database.ads
@@ -90,6 +90,9 @@ package Podmander.Database is
    function Column_Int (QH : Query_Handle; Index : Natural) return Integer;
    -- Read an integer column from the current row.
 
+   function Column_Is_Null (QH : Query_Handle; Index : Natural) return Boolean;
+   -- Return True if the column at Index is NULL in the current row.
+
    function Changes (DB : in out DB_Handle) return Integer;
    -- Rows modified by most recent INSERT/UPDATE/DELETE.